[BugFix]WeightCompositの戻り値をunsigendからdoubleに変更

+++ Hal/Heuristics/WeightComposit.cpp
@@ -1,13 +1,13 @@
 
 #include "WeightComposit.h"
 
-void WeightComposit::AddHeuristic(h_type* heuristics,float weight){
-	h_list.insert(std::pair<h_type*,float>(heuristics,weight));
+void WeightComposit::AddHeuristic(h_type* heuristics,double weight){
+	h_list.insert(std::pair<h_type*,double>(heuristics,weight));
 }
 
-unsigned WeightComposit::Execution(const Field& field){
-	unsigned score = 0;
-	for(std::pair<h_type*,float> h : h_list){
+double WeightComposit::Execution(const Field& field){
+	double score = 0;
+	for(std::pair<h_type*,double> h : h_list){
 		score += h.first->Execution(field) * h.second;
 	}
 	return score;
